+++ module.json
@@ -1,37 +1,41 @@
 {
   "id": "find-the-culprit",
   "name": "find-the-culprit",
-  "title": "Find the culprit",
+  "title": "Find the Culprit",
   "description": "This module helps you to find a module causing issues, by automating the process of deactivating and reactivating modules.<br>Just click the 'Find the culprit' button inside Module Management.",
-  "author": "Moerill",
   "authors": [
     {
       "name": "Moerill"
+    },
+    {
+      "name": "Emmanuel Wineberg",
+      "email": "esheyw@gmail.com",
+      "url": "https://github.com/esheyw",
+      "discord": "esheyw"
     }
   ],
-  "url": "https://github.com/Moerill/fvtt-find-the-culprit",
-  "license": "https://github.com/Moerill/fvtt-find-the-culprit/blob/master/LICENSE",
-  "readme": "https://github.com/Moerill/fvtt-find-the-culprit#find-the-culprit",
-  "bugs": "https://github.com/Moerill/fvtt-find-the-culprit/issues",
-  "changelog": "https://github.com/Moerill/fvtt-find-the-culprit/blob/master/CHANGELOG.md",
+  "url": "https://github.com/eshey/find-the-culprit",
+  "license": "https://github.com/eshey/find-the-culprit/blob/master/LICENSE",
+  "readme": "https://github.com/eshey/find-the-culprit#find-the-culprit",
+  "bugs": "https://github.com/eshey/find-the-culprit/issues",
+  "changelog": "https://github.com/eshey/find-the-culprit/blob/master/CHANGELOG.md",
   "flags": {},
-  "version": "1.4.6",
+  "version": "1.5.0",
   "compatibility": {
-    "minimum": "9",
-    "verified": "10"
+    "minimum": "11",
+    "verified": "12"
   },
-  "minimumCoreVersion": "9",
-  "compatibleCoreVersion": "10",
-  "scripts": [],
-  "esmodules": ["js/find-the-culprit.js"],
-  "styles": ["css/find-the-culprit.css"],
-  "languages": [],
-  "packs": [],
-  "system": [],
+  "esmodules": ["scripts/find-the-culprit.mjs"],
+  "styles": ["styles/find-the-culprit.css"],
+  "languages": [
+    {
+      "lang": "en",
+      "name": "English",
+      "path": "lang/en.json",
+      "flags": {}
+    }
+  ],
   "relationships": {},
-  "dependencies": [],
-  "socket": false,
-  "manifest": "https://raw.githubusercontent.com/Moerill/fvtt-find-the-culprit/master/module.json",
-  "download": "https://github.com/Moerill/fvtt-find-the-culprit/releases/download/v1.4.6/v1.4.6.zip",
-  "protected": false
+  "manifest": "{{updatedByRelease}}",
+  "download": "{{updatedByRelease}}"
 }
