+++ templates/dependenciesDialog.hbs
@@ -1,0 +1,17 @@
+<form>   
+  {{#if enabling}}
+    <p>{{ localize "MODMANAGE.DepRequiredEnable" number=dependencies.length }}</p>
+    {{#each dependencies as |dependency|}}
+    <div class="form-group">
+      <label class="checkbox"><input type="checkbox" name="{{dependency.id}}" checked="checked" data-tooltip="{{ dependency.reason }}">{{dependency.title}}</label>
+    </div>
+    {{/each}} 
+  {{else}}
+    <p>{{ localize "MODMANAGE.DepDisable" number=dependencies.length }}</p> 
+    {{#each dependencies as |dependency|}}
+    <div class="form-group">
+      <label class="checkbox"><input type="checkbox" name="{{dependency.id}}" checked="checked" data-tooltip="{{ dependency.reason }}">{{dependency.title}}</label>
+    </div>
+    {{/each}}
+  {{/if}}
+ </form>
