+++ templates/foundTheCulprit.hbs
@@ -1,0 +1,4 @@
+<h2>{{localize "FindTheCulprit.FoundTheCulprit.Header"}}</h2>
+<ul class="ftc-module-list">
+  <li title="{{localize "FindTheCulprit.CurrentlyActive"}}"><i class="fa-solid fa-check ftc-active"></i>{{culpritTitle}}</li>
+</ul>
