+++ templates/selectMods.hbs
@@ -1,0 +1,17 @@
+<p>{{localize "FindTheCulprit.SelectMods.Prompt"}}</p>
+<input type="text" name="search" placeholder="Filter Modules" value="">
+<ul class='ftc-module-list ftc-module-chooser'>
+  {{#each activeModules}}
+  <li data-module="{{id}}">
+    <label class="ftc-lock-checkbox">
+      <input class="lock-btn hidden" type="checkbox" data-module="{{id}}" tabindex="-1" {{checked (lookup @root.locks id)}}/>
+      <span class="fa-solid lock"></span>
+    </label>
+    <input class="ftc-checkbox" type="checkbox" data-module="{{id}}" id="find-the-culprit-{{id}}" {{checked (lookup @root.locks id)}}/>
+    <label class="package-title" for="find-the-culprit-{{id}}">{{title}}</label>
+  </li>
+  {{/each}}              
+</ul>
+<p>{{localize "FindTheCulprit.SelectMods.Usage.P1"}}</p>
+<p>{{localize "FindTheCulprit.SelectMods.Usage.P2"}}</p>
+<p>{{localize "FindTheCulprit.SelectMods.Usage.P2"}}</p>
